python-common/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "campusvista-recognition"
version = "0.1.0"
description = "Shared CampusVista visual recognition encoders and image preprocessing helpers."
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26,<3.0",
    "Pillow>=10.0,<12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["campusvista_recognition*"]
